requirements.txt
# Connect 4 AI - Requirements for Hugging Face Deployment
# Python 3.7+

# Core dependencies for web interface
gradio>=4.0.0
Pillow>=9.0.0
numpy>=1.21.0

# Optional: Development dependencies (uncomment if needed locally)
# pytest>=7.0.0  # For running tests
# black>=22.0.0  # For code formatting
# flake8>=4.0.0  # For linting
